refactor(vm): Refactor instruction handling
Move instruction handling to its own module. This will make it easier to
add new instructions in the future.
Make the execution of instructions more generic by using enums to
represent the instructions.
