+++ vm/src/component/mod.rs
@@ -1,5 +1,6 @@
 pub mod cpu;
 pub mod device;
+pub mod instructions;
 pub mod memory;
 pub mod memory_io;
 pub mod memory_map;
